backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==4.2.6
django-elasticsearch-dsl==8.0
django-extensions==3.2.0
djangorestframework==3.14.0
elastic-transport==8.17.1
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
Faker==37.5.3
graphviz==0.20.1
psycopg2-binary==2.9.6
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
